ocr/requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
paddleocr==2.6.0.1
paddlepaddle==2.4.2  # Utilisez paddlepaddle==2.4.2 si vous utilisez la version CPU
Pillow==9.2.0
numpy==1.22.4
python-multipart==0.0.5
PyMuPDF==1.22.3  # Pour l'extraction de texte à partir de PDF
pdf2image==1.16.3  # Pour la conversion des pages PDF en images
